WIP: 910e22d Merge pull request #305 from EmperorArthur/gridfinity-rebuilt-2
Adds magnet hole placement options

Adds options to place magnet holes on the top or bottom of the Gridfinity baseplate.

This allows for more flexible configurations, enabling magnets to be placed on either or both surfaces of the baseplate. Includes updates to the baseplate module to handle the new parameters and hole creation logic.